refactor & add update/delete posts

+++ components/post/action-button.tsx
@@ -1,10 +1,11 @@
 "use client"
 import { updatePostStatus } from "@/actions/posts/update-post-status"
 import { useState } from "react"
-import { BiLoader } from "react-icons/bi"
+import { BiLoader, BiX } from "react-icons/bi"
 import { toast } from "sonner"
 import { Button } from "../ui/button"
 import { deletePost } from "@/actions/posts/delete-post"
+import { CheckIcon, Cross1Icon, TrashIcon } from "@radix-ui/react-icons"
 
 export const ActionButton = ({
   postId,
@@ -13,7 +14,7 @@ export const ActionButton = ({
 }: {
   postId: string
   type: "APPROVE" | "REJECT" | "DELETE"
-  onActionSuccess: (postId: string) => void
+  onActionSuccess?: (postId: string) => void
 }) => {
   const [loading, setLoading] = useState(false)
 
@@ -29,12 +30,19 @@ export const ActionButton = ({
       return toast.error(res.error)
     }
 
-    onActionSuccess(postId)
+    onActionSuccess?.(postId)
     return toast.success(res.success)
   }
 
   const handleDelete = async () => {
+    const confirmation = prompt(
+      "Deseja realmente deletar este anúncio? (digite 'sim')"
+    )
+
+    if (!confirmation || confirmation.toLocaleLowerCase() !== "sim") return
+
     setLoading(true)
+
     const res = await deletePost(postId)
 
     setLoading(false)
@@ -43,7 +51,7 @@ export const ActionButton = ({
       return toast.error(res.error)
     }
 
-    onActionSuccess(postId)
+    onActionSuccess?.(postId)
     return toast.success(res.success)
   }
 
@@ -67,7 +75,7 @@ export const ActionButton = ({
       return toast.error(res.error)
     }
 
-    onActionSuccess(postId)
+    onActionSuccess?.(postId)
     return toast.success(res.success)
   }
 
@@ -75,10 +83,14 @@ export const ActionButton = ({
     return (
       <Button
         onClick={handleDelete}
-        className="min-w-[85px]"
         aria-label={loading ? "deletando anúncio..." : "deletar anúncio?"}
+        size="icon"
       >
-        {loading ? <BiLoader className="animate-spin" /> : "Deletar"}
+        {loading ? (
+          <BiLoader className="animate-spin" />
+        ) : (
+          <TrashIcon fontSize={24} />
+        )}
       </Button>
     )
   }
@@ -86,22 +98,32 @@ export const ActionButton = ({
   if (type === "REJECT") {
     return (
       <Button
-        className="bg-red-500 hover:bg-red-600 min-w-[85px]"
+        className="bg-red-500 hover:bg-red-600"
         onClick={handleReject}
         aria-label={loading ? "rejeitando anúncio..." : "rejeitar anúncio?"}
+        size="icon"
       >
-        {loading ? <BiLoader className="animate-spin" /> : "Rejeitar"}
+        {loading ? (
+          <BiLoader className="animate-spin" />
+        ) : (
+          <Cross1Icon fontSize={20} />
+        )}
       </Button>
     )
   }
 
   return (
     <Button
-      className="bg-green-500 hover:bg-green-600 min-w-[85px]"
+      className="bg-green-500 hover:bg-green-600"
       onClick={handleApprove}
+      size="icon"
       aria-label={loading ? "aprovando anúncio..." : "aprovar anúncio?"}
     >
-      {loading ? <BiLoader className="animate-spin" /> : "Aprovar"}
+      {loading ? (
+        <BiLoader className="animate-spin" />
+      ) : (
+        <CheckIcon fontSize={24} />
+      )}
     </Button>
   )
 }
